+++ lib/attesto_phoenix/store/par_ets.ex
@@ -19,6 +19,24 @@ defmodule AttestoPhoenix.Store.PAR.ETS do
   end
 
   @impl true
+  def fetch(request_uri) when is_binary(request_uri) do
+    ensure_table()
+    now = System.system_time(:second)
+
+    case :ets.lookup(@table, request_uri) do
+      [{^request_uri, params, expires_at}] when expires_at > now ->
+        {:ok, params}
+
+      [{^request_uri, _params, _expires_at}] ->
+        :ets.delete(@table, request_uri)
+        :error
+
+      [] ->
+        :error
+    end
+  end
+
+  @impl true
   def take(request_uri) when is_binary(request_uri) do
     ensure_table()
     now = System.system_time(:second)